requirements.txt
# Dependencias para Azure Functions
azure-functions
azure-identity
azure-ai-textanalytics
azure-ai-contentsafety
azure-core

# OpenAI en Azure
azure-openai @ git+https://github.com/Azure/azure-sdk-for-python.git#subdirectory=sdk/openai/azure-openai
